feat: 회원 DTO를 프로필 DTO로 대체
