+++ backend/src/main/java/com/woowacourse/f12/dto/request/profile/ProfileSearchRequest.java
@@ -1,21 +1,21 @@
-package com.woowacourse.f12.dto.request.member;
+package com.woowacourse.f12.dto.request.profile;
 
 import com.woowacourse.f12.presentation.member.CareerLevelConstant;
 import com.woowacourse.f12.presentation.member.JobTypeConstant;
 import lombok.Getter;
 
 @Getter
-public class MemberSearchRequest {
+public class ProfileSearchRequest {
 
     private String query;
     private CareerLevelConstant careerLevel;
     private JobTypeConstant jobType;
 
-    private MemberSearchRequest() {
+    private ProfileSearchRequest() {
     }
 
-    public MemberSearchRequest(final String query, final CareerLevelConstant careerLevel,
-                               final JobTypeConstant jobType) {
+    public ProfileSearchRequest(final String query, final CareerLevelConstant careerLevel,
+                                final JobTypeConstant jobType) {
         this.query = query;
         this.careerLevel = careerLevel;
         this.jobType = jobType;
